+++ CHANGELOG.md
@@ -19,6 +19,7 @@
 - Search within community
 - OP indicator in the comments
 - Post reporting
+- Comment reporting
 
 ### Changes